requirements.txt
rumps>=0.4.0
pyaudio>=0.2.14
numpy>=1.24.0
mlx-whisper>=0.4.0
pyperclip>=1.8.0
pyautogui>=0.9.0
pynput>=1.7.0
pyobjc-framework-Cocoa>=10.0
requests>=2.31.0
